refactor[#118]: 리팩토링
