+++ src/main/java/org/highfive/backend/content/controller/OnboardingController.java
@@ -4,14 +4,15 @@ import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.highfive.backend.auth.dto.response.TokenResponseDto;
 import org.highfive.backend.content.dto.response.OnboardingInitContentsResponseDto;
 import org.highfive.backend.content.dto.request.OnboardingSelectContentRequestDto;
 import org.highfive.backend.content.dto.response.OnboardingSelectContentResponseDto;
-import org.highfive.backend.content.service.ContentService;
 import org.highfive.backend.content.service.OnboardingService;
-import org.highfive.backend.global.code.SuccessCode;
 import org.highfive.backend.global.dto.Response;
+import org.highfive.backend.user.dto.request.SubmitOnboardingRequestDto;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -21,27 +22,20 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class OnboardingController {
 
-    private final ContentService contentService;
     private final OnboardingService onboardingService;
 
-    /**
-     * 온보딩 초기 6개의 작품을 띄웁니다.
-     *
-     * @return
-     */
     @GetMapping("/content/init")
     public Response<List<OnboardingInitContentsResponseDto>> onboardingInitContents() {
-        return new Response<>(SuccessCode.OK.getCode(), contentService.getDistinctGenreTopContents(), null);
+        return onboardingService.getDistinctGenreTopContents();
     }
 
     @PostMapping("/content/recommend")
     public Response<List<OnboardingSelectContentResponseDto>> selectContent(@Valid @RequestBody final OnboardingSelectContentRequestDto request) {
-        final List<OnboardingSelectContentResponseDto> contents = onboardingService.getContentBySelectedContent(request);
+        return onboardingService.getContentBySelectedContent(request);
+    }
 
-        if (contents.isEmpty()) {
-            return new Response<>(SuccessCode.NO_CONTENT.getCode(), contents, null);
-        }
-
-        return new Response<>(SuccessCode.OK.getCode(), contents, null);
+    @PatchMapping("/user/info")
+    public Response<TokenResponseDto> submitOnboarding(@Valid @RequestBody final SubmitOnboardingRequestDto request) {
+        return onboardingService.initUser(request);
     }
 }
