+++ src/test/java/org/highfive/backend/content/service/ContentServiceTest.java
@@ -56,85 +56,8 @@ public class ContentServiceTest {
     @Mock
     private MetaInfoContentsRepository metaInfoContentsRepository;
 
-    @Mock
-    private FastApiClient fastApiClient;
-
     @InjectMocks
     private ContentService contentService;
-    @Mock
-    private PreferMetaInfoRepository preferMetaInfoRepository;
-
-    private List<MostPopularContentPerGenreDto> stubData;
-
-    @BeforeEach
-    void setUp() {
-        stubData = List.of(
-                new MostPopularContentPerGenreDto(1L, "s3://1", "A의 모험", 10, "장르1", 1L, 2024),
-                new MostPopularContentPerGenreDto(2L, "s3://2", "B의 비밀", 10, "장르2", 2L, 2024)
-        );
-    }
-
-    @DisplayName("온보딩 초기 - ContentRepository의 리턴 타입이 ContentService의 리턴 타입으로 잘 매핑된다")
-    @Test
-    void getDistinctGenreTopContents_success() {
-        // given
-        when(contentRepository.findTopContentPerGenre(6)).thenReturn(stubData);
-
-        // when
-        List<OnboardingInitContentsResponseDto> result = contentService.getDistinctGenreTopContents();
-
-        // then
-        assertThat(result)
-                .hasSize(2)
-                .extracting("contentId", "thumbnailUrl", "title")
-                .containsExactly(
-                        tuple(1L, "s3://1", "A의 모험"),
-                        tuple(2L, "s3://2", "B의 비밀")
-                );
-
-        verify(contentRepository).findTopContentPerGenre(6);
-        verifyNoMoreInteractions(contentRepository, metaInfoContentsRepository,
-                fastApiClient, preferMetaInfoRepository);
-    }
-
-    @Test
-    @DisplayName("온보딩 초기 - 장르별 인기있는 작품 리스트가 빈 리스트인 경우 예외가 발생한다")
-    void getDistinctGenreTopContents_emptyResult() {
-        // given
-        when(contentRepository.findTopContentPerGenre(6))
-                .thenReturn(List.of());
-
-        // when
-        assertThatThrownBy(() -> contentService.getDistinctGenreTopContents())
-                .isInstanceOf(BusinessException.class)
-                .satisfies(ex -> {
-                    BusinessException be = (BusinessException) ex;
-                    assertThat(be.getErrorCode()).isEqualTo(ContentErrorCode.CONTENT_NOT_FOUND);
-                });
-
-        // then
-        verify(contentRepository).findTopContentPerGenre(6);
-        verifyNoMoreInteractions(contentRepository, metaInfoContentsRepository,
-                fastApiClient, preferMetaInfoRepository);
-    }
-
-    @Test
-    @DisplayName("온보딩 초기 - ContentRepository가 예외를 던지면 서비스도 예외를 전파한다")
-    void getDistinctGenreTopContents_repositoryThrows() {
-        // given
-        when(contentRepository.findTopContentPerGenre(6))
-                .thenThrow(new IllegalStateException("DB 오류"));
-
-        // when, then
-        assertThatThrownBy(() -> contentService.getDistinctGenreTopContents())
-                .isInstanceOf(IllegalStateException.class)
-                .hasMessageContaining("DB 오류");
-
-        verify(contentRepository).findTopContentPerGenre(6);
-        verifyNoMoreInteractions(contentRepository, metaInfoContentsRepository,
-                fastApiClient, preferMetaInfoRepository);
-    }
-
 
     @Test
     @DisplayName("존재하는 content일 경우 content의 상세 정보를 반환한다")
@@ -189,64 +112,5 @@ public class ContentServiceTest {
 
         //then
         assertEquals(ContentErrorCode.CONTENT_NOT_FOUND, exception.getErrorCode());
-    }
-
-    @Nested
-    @DisplayName("홈 화면 메인 컨텐츠")
-    class RecommendMain {
-
-        @Test
-        @DisplayName("dto 매핑이 정상적으로 이루어집니다")
-        void success() {
-            given(fastApiClient.getContentsByVector(anyString(), eq(1))).willReturn(List.of(new FastApiRecommendResponseDto(1L)));
-
-            Content c = ContentFixture.createHomeMainContent(1L, "poster.jpg", "desc");
-            given(contentRepository.findById(1L)).willReturn(Optional.of(c));
-            given(contentRepository.findContentGenresByContentIds(List.of(1L))).willReturn(List.of(Map.of("genreName", "Action")));
-
-            MainRecommendDto dto = contentService.recommendMainContentsByUser(UserFixture.createEmbeddingUser());
-
-            assertThat(dto.posterUrl()).isEqualTo("poster.jpg");
-            assertThat(dto.description()).isEqualTo("desc");
-            assertThat(dto.genre()).containsExactly("Action");
-        }
-
-        @Test
-        @DisplayName("추천 컨텐츠가 없는 경우 CONTENT_NOT_FOUND 예외를 발생시킵니다")
-        void notFound() {
-            given(fastApiClient.getContentsByVector(anyString(), eq(1))).willReturn(List.of(new FastApiRecommendResponseDto(1L)));
-            given(contentRepository.findById(1L)).willReturn(java.util.Optional.empty());
-
-            BusinessException ex = assertThrows(BusinessException.class, () -> contentService.recommendMainContentsByUser(UserFixture.createEmbeddingUser()));
-
-            assertThat(ex.getErrorCode()).isEqualTo(ContentErrorCode.CONTENT_NOT_FOUND);
-        }
-    }
-
-    @Nested
-    @DisplayName("개인화 추천 컨텐츠 조회")
-    class RecommendPersonal {
-
-        @Test
-        @DisplayName("dto 매핑이 정상적으로 이루어집니다")
-        void success() {
-            List<FastApiRecommendResponseDto> fastDtos = List.of(
-                    new FastApiRecommendResponseDto(200L),
-                    new FastApiRecommendResponseDto(201L),
-                    new FastApiRecommendResponseDto(202L)
-            );
-            given(fastApiClient.getContentsByVector(anyString(), eq(3))).willReturn(fastDtos);
-
-            for (long id : List.of(200L, 201L, 202L)) {
-                Content c = ContentFixture.createContent(id);
-                given(contentRepository.findById(id)).willReturn(java.util.Optional.of(c));
-            }
-
-            List<PersonalRecommendDto> list = contentService.recommendContentsByUser(UserFixture.createEmbeddingUser(), 3);
-
-            assertThat(list).hasSize(3)
-                    .extracting("contentId")
-                    .containsExactly(200L, 201L, 202L);
-        }
     }
 }
