+++ src/test/java/org/highfive/backend/content/service/OnboardingServiceTest.java
@@ -1,16 +1,30 @@
 package org.highfive.backend.content.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForClassTypes.tuple;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import org.assertj.core.api.AssertionsForClassTypes;
+import org.assertj.core.api.AssertionsForInterfaceTypes;
 import org.highfive.backend.content.dto.request.OnboardingSelectContentRequestDto;
 import org.highfive.backend.content.dto.response.GenreCountDto;
+import org.highfive.backend.content.dto.response.MostPopularContentPerGenreDto;
 import org.highfive.backend.content.dto.response.OnboardingContentDto;
+import org.highfive.backend.content.dto.response.OnboardingInitContentsResponseDto;
 import org.highfive.backend.content.dto.response.OnboardingSelectContentResponseDto;
+import org.highfive.backend.content.entity.repository.MetaInfoContentsRepository;
+import org.highfive.backend.content.exception.ContentErrorCode;
+import org.highfive.backend.content.repository.ContentRepository;
 import org.highfive.backend.content.repository.QueryDslContentRepository;
+import org.highfive.backend.global.client.fastapi.FastApiClient;
+import org.highfive.backend.global.dto.Response;
+import org.highfive.backend.global.exception.BusinessException;
+import org.highfive.backend.user.entity.preference.PreferMetaInfoRepository;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -26,6 +40,28 @@ class OnboardingServiceTest {
 
     @InjectMocks
     OnboardingService onboardingService;
+
+    @Mock
+    private MetaInfoContentsRepository metaInfoContentsRepository;
+
+    @Mock
+    private PreferMetaInfoRepository preferMetaInfoRepository;
+
+    @Mock
+    private ContentRepository contentRepository;
+
+    @Mock
+    private FastApiClient fastApiClient;
+
+    private List<MostPopularContentPerGenreDto> stubData;
+
+    @BeforeEach
+    void setUp() {
+        stubData = List.of(
+                new MostPopularContentPerGenreDto(1L, "s3://1", "A의 모험", 10, "장르1", 1L, 2024),
+                new MostPopularContentPerGenreDto(2L, "s3://2", "B의 비밀", 10, "장르2", 2L, 2024)
+        );
+    }
 
     @Test
     @DisplayName("온보딩 선택 - 선택한 콘텐츠 ID들을 기반으로 중복을 제외한 3개 이하의 추천 콘텐츠를 반환한다")
@@ -53,10 +89,10 @@ class OnboardingServiceTest {
                 .thenReturn(repoReturn);
 
         // when
-        List<OnboardingSelectContentResponseDto> actual = onboardingService.getContentBySelectedContent(req);
+        Response<List<OnboardingSelectContentResponseDto>> actual = onboardingService.getContentBySelectedContent(req);
 
         // then
-        assertThat(actual).containsExactly(
+        assertThat(actual.content()).containsExactly(
                         new OnboardingSelectContentResponseDto(3L, "url3", "title3", 2022),
                         new OnboardingSelectContentResponseDto(4L, "url4", "title4", 2021),
                         new OnboardingSelectContentResponseDto(5L, "url5", "title5", 2020)
@@ -65,5 +101,66 @@ class OnboardingServiceTest {
         verify(queryDslRepo).findTopGenresByContentIds(selectedIds);
         verify(queryDslRepo).findContentsByGenresOrderByMatchCountDesc(List.of("Action", "Drama"));
         verifyNoMoreInteractions(queryDslRepo);
+    }
+
+    @DisplayName("온보딩 초기 - ContentRepository의 리턴 타입이 ContentService의 리턴 타입으로 잘 매핑된다")
+    @Test
+    void getDistinctGenreTopContents_success() {
+        // given
+        when(contentRepository.findTopContentPerGenre(6)).thenReturn(stubData);
+
+        // when
+        Response<List<OnboardingInitContentsResponseDto>> result = onboardingService.getDistinctGenreTopContents();
+
+        // then
+        AssertionsForInterfaceTypes.assertThat(result.content())
+                .hasSize(2)
+                .extracting("contentId", "thumbnailUrl", "title")
+                .containsExactly(
+                        tuple(1L, "s3://1", "A의 모험"),
+                        tuple(2L, "s3://2", "B의 비밀")
+                );
+
+        verify(contentRepository).findTopContentPerGenre(6);
+        verifyNoMoreInteractions(contentRepository, metaInfoContentsRepository,
+                fastApiClient, preferMetaInfoRepository);
+    }
+
+    @Test
+    @DisplayName("온보딩 초기 - 장르별 인기있는 작품 리스트가 빈 리스트인 경우 예외가 발생한다")
+    void getDistinctGenreTopContents_emptyResult() {
+        // given
+        when(contentRepository.findTopContentPerGenre(6))
+                .thenReturn(List.of());
+
+        // when
+        assertThatThrownBy(() -> onboardingService.getDistinctGenreTopContents())
+                .isInstanceOf(BusinessException.class)
+                .satisfies(ex -> {
+                    BusinessException be = (BusinessException) ex;
+                    AssertionsForClassTypes.assertThat(be.getErrorCode()).isEqualTo(ContentErrorCode.CONTENT_NOT_FOUND);
+                });
+
+        // then
+        verify(contentRepository).findTopContentPerGenre(6);
+        verifyNoMoreInteractions(contentRepository, metaInfoContentsRepository,
+                fastApiClient, preferMetaInfoRepository);
+    }
+
+    @Test
+    @DisplayName("온보딩 초기 - ContentRepository가 예외를 던지면 서비스도 예외를 전파한다")
+    void getDistinctGenreTopContents_repositoryThrows() {
+        // given
+        when(contentRepository.findTopContentPerGenre(6))
+                .thenThrow(new IllegalStateException("DB 오류"));
+
+        // when, then
+        assertThatThrownBy(() -> onboardingService.getDistinctGenreTopContents())
+                .isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("DB 오류");
+
+        verify(contentRepository).findTopContentPerGenre(6);
+        verifyNoMoreInteractions(contentRepository, metaInfoContentsRepository,
+                fastApiClient, preferMetaInfoRepository);
     }
 }
